+++ src/braket/circuits/angled_gate.py
@@ -232,6 +232,133 @@ class DoubleAngledGate(Gate, Parameterizable):
         )
 
 
+class TripleAngledGate(Gate, Parameterizable):
+    """
+    Class `TripleAngledGate` represents a quantum gate that operates on N qubits and three angles.
+    """
+
+    def __init__(
+        self,
+        angle_1: Union[FreeParameterExpression, float],
+        angle_2: Union[FreeParameterExpression, float],
+        angle_3: Union[FreeParameterExpression, float],
+        qubit_count: Optional[int],
+        ascii_symbols: Sequence[str],
+    ):
+        """
+        Args:
+            angle_1 (Union[FreeParameterExpression, float]): The first angle of the gate in
+                radians or expression representation.
+            angle_2 (Union[FreeParameterExpression, float]): The second angle of the gate in
+                radians or expression representation.
+            angle_3 (Union[FreeParameterExpression, float]): The third angle of the gate in
+                radians or expression representation.
+            qubit_count (Optional[int]): The number of qubits that this gate interacts with.
+            ascii_symbols (Sequence[str]): ASCII string symbols for the gate. These are used when
+                printing a diagram of a circuit. The length must be the same as `qubit_count`, and
+                index ordering is expected to correlate with the target ordering on the instruction.
+                For instance, if a CNOT instruction has the control qubit on the first index and
+                target qubit on the second index, the ASCII symbols should have `["C", "X"]` to
+                correlate a symbol with that index.
+
+        Raises:
+            ValueError: If `qubit_count` is less than 1, `ascii_symbols` are `None`, or
+                `ascii_symbols` length != `qubit_count`, or `angle_1` or `angle_2` or `angle_3`
+                 is `None`
+        """
+        super().__init__(qubit_count=qubit_count, ascii_symbols=ascii_symbols)
+        if angle_1 is None or angle_2 is None or angle_3 is None:
+            raise ValueError("angles must not be None")
+        self._parameters = [
+            (
+                angle
+                if isinstance(angle, FreeParameterExpression)
+                else float(angle)  # explicit casting in case angle is e.g. np.float32
+            )
+            for angle in (angle_1, angle_2, angle_3)
+        ]
+
+    @property
+    def parameters(self) -> List[Union[FreeParameterExpression, float]]:
+        """
+        Returns the parameters associated with the object, either unbound free parameters or
+        bound values.
+
+        Returns:
+            List[Union[FreeParameterExpression, float]]: The free parameters or fixed value
+            associated with the object.
+        """
+        return self._parameters
+
+    @property
+    def angle_1(self) -> Union[FreeParameterExpression, float]:
+        """
+        Returns the first angle for the gate
+
+        Returns:
+            Union[FreeParameterExpression, float]: The first angle of the gate in radians
+        """
+        return self._parameters[0]
+
+    @property
+    def angle_2(self) -> Union[FreeParameterExpression, float]:
+        """
+        Returns the second angle for the gate
+
+        Returns:
+            Union[FreeParameterExpression, float]: The second angle of the gate in radians
+        """
+        return self._parameters[1]
+
+    @property
+    def angle_3(self) -> Union[FreeParameterExpression, float]:
+        """
+        Returns the second angle for the gate
+
+        Returns:
+            Union[FreeParameterExpression, float]: The third angle of the gate in radians
+        """
+        return self._parameters[2]
+
+    def bind_values(self, **kwargs) -> AngledGate:
+        """
+        Takes in parameters and attempts to assign them to values.
+
+        Args:
+            ``**kwargs``: The parameters that are being assigned.
+
+        Returns:
+            AngledGate: A new Gate of the same type with the requested parameters bound.
+
+        Raises:
+            NotImplementedError: Subclasses should implement this function.
+        """
+        raise NotImplementedError
+
+    def adjoint(self) -> List[Gate]:
+        """Returns the adjoint of this gate as a singleton list.
+
+        Returns:
+            List[Gate]: A list containing the gate with negated angle.
+        """
+        raise NotImplementedError
+
+    def __eq__(self, other):
+        return (
+            isinstance(other, TripleAngledGate)
+            and self.name == other.name
+            and _angles_equal(self.angle_1, other.angle_1)
+            and _angles_equal(self.angle_2, other.angle_2)
+            and _angles_equal(self.angle_3, other.angle_3)
+        )
+
+    def __repr__(self):
+        return (
+            f"{self.name}('angles': ({self.angle_1}, {self.angle_2}, {self.angle_3}), "
+            f"'qubit_count': {self.qubit_count})"
+        )
+
+
 @singledispatch
 def _angles_equal(
     angle_1: Union[FreeParameterExpression, float], angle_2: Union[FreeParameterExpression, float]
@@ -259,28 +386,26 @@ def angled_ascii_characters(gate: str, angle: Union[FreeParameterExpression, flo
     return f'{gate}({angle:{".2f" if isinstance(angle, (float, Float)) else ""}})'
 
 
-def _double_angled_ascii_characters(
+def _multi_angled_ascii_characters(
     gate: str,
-    angle_1: Union[FreeParameterExpression, float],
-    angle_2: Union[FreeParameterExpression, float],
+    *angles: Union[FreeParameterExpression, float],
 ) -> str:
     """
     Generates a formatted ascii representation of an angled gate.
 
     Args:
         gate (str): The name of the gate.
-        angle_1 (Union[FreeParameterExpression, float]): angle in radians.
-        angle_2 (Union[FreeParameterExpression, float]): angle in radians.
+        `*angles` (Union[FreeParameterExpression, float]): angles in radians.
 
     Returns:
         str: Returns the ascii representation for an angled gate.
 
     """
-    return (
-        f"{gate}("
-        f'{angle_1:{".2f" if isinstance(angle_1, (float, Float)) else ""}}, '
-        f'{angle_2:{".2f" if isinstance(angle_2, (float, Float)) else ""}})'
-    )
+
+    def format_string(angle: Union[FreeParameterExpression, float]) -> str:
+        return ".2f" if isinstance(angle, (float, Float)) else ""
+
+    return f"{gate}({', '.join(f'{angle:{format_string(angle)}}' for angle in angles)})"
 
 
 def get_angle(gate: AngledGate, **kwargs) -> AngledGate:
@@ -301,17 +426,17 @@ def get_angle(gate: AngledGate, **kwargs) -> AngledGate:
     return type(gate)(angle=new_angle)
 
 
-def _get_angles(gate: DoubleAngledGate, **kwargs) -> DoubleAngledGate:
+def _get_angles(gate: TripleAngledGate, **kwargs) -> TripleAngledGate:
     """
     Gets the angle with all values substituted in that are requested.
 
     Args:
-        gate (DoubleAngledGate): The subclass of DoubleAngledGate for which the angle is being
+        gate (TripleAngledGate): The subclass of TripleAngledGate for which the angle is being
             obtained.
         ``**kwargs``: The named parameters that are being filled for a particular gate.
 
     Returns:
-        DoubleAngledGate: A new gate of the type of the AngledGate originally used with all angles
+        TripleAngledGate: A new gate of the type of the AngledGate originally used with all angles
         updated.
     """
     new_angles = [
@@ -320,6 +445,6 @@ def _get_angles(gate: DoubleAngledGate, **kwargs) -> DoubleAngledGate:
             if isinstance(getattr(gate, angle), FreeParameterExpression)
             else getattr(gate, angle)
         )
-        for angle in ("angle_1", "angle_2")
+        for angle in ("angle_1", "angle_2", "angle_3")
     ]
-    return type(gate)(angle_1=new_angles[0], angle_2=new_angles[1])
+    return type(gate)(angle_1=new_angles[0], angle_2=new_angles[1], angle_3=new_angles[2])