+++ src/braket/devices/__init__.py
@@ -12,4 +12,5 @@
 # language governing permissions and limitations under the License.
 
 from braket.devices.device import Device  # noqa: F401
+from braket.devices.devices import Devices  # noqa: F401
 from braket.devices.local_simulator import LocalSimulator  # noqa: F401